services/consumer/requirements.txt
google-generativeai
prometheus_client==0.21.1
python-dateutil==2.9.0.post0
